# The ps_trace_thread does not reset the settings for the performance schema when the "File '%s' already exists" exception occurs.