The ps_trace_thread does not reset the settings for the performance schema when the "File '%s' already exists" exception occurs.
 
The ps_trace_thread procedure does not handle the "File '%s' already exists"  exception.
Thereby when the file specified as the "in_outfile" parameter argument exists before the ps_trace_thread execution, the unhandled "File '%s' already exists" exception causes the ps_trace_thread not to reset the settings for the performance schema. 

How to reproduce:
```
mysql> SELECT COUNT(*) FROM performance_schema.setup_actors;
+----------+
| COUNT(*) |
+----------+
|        2 |
+----------+
1 row in set (0.05 sec)

mysql> SELECT COUNT(*) FROM performance_schema.threads WHERE INSTRUMENTED = 'YES';
+----------+
| COUNT(*) |
+----------+
|       35 |
+----------+
1 row in set (0.05 sec)

mysql> CALL sys.ps_trace_thread(44, '/var/lib/mysql-files/stack1.dot', 2, 1, TRUE, TRUE, TRUE);
+-------------------+
| summary           |
+-------------------+
| Disabled 1 thread |
+-------------------+
1 row in set (0.05 sec)

+---------------------------------------------+
| Info                                        |
+---------------------------------------------+
| Data collection starting for THREAD_ID = 44 |
+---------------------------------------------+
1 row in set (0.09 sec)

ERROR 1086 (HY000): File '/var/lib/mysql-files/stack1.dot' already exists
mysql> SELECT COUNT(*) FROM performance_schema.setup_actors;
+----------+
| COUNT(*) |
+----------+
|        0 |
+----------+
1 row in set (0.05 sec)

mysql> SELECT COUNT(*) FROM performance_schema.threads WHERE INSTRUMENTED = 'YES';
+----------+
| COUNT(*) |
+----------+
|        1 |
+----------+
1 row in set (0.05 sec)
```

After fix:
```
mysql> SELECT COUNT(*) FROM performance_schema.setup_actors;
+----------+
| COUNT(*) |
+----------+
|        2 |
+----------+
1 row in set (0.04 sec)

mysql> SELECT COUNT(*) FROM performance_schema.threads WHERE INSTRUMENTED = 'YES';
+----------+
| COUNT(*) |
+----------+
|       34 |
+----------+
1 row in set (0.05 sec)

mysql> CALL sys.ps_trace_thread(75, '/var/lib/mysql-files/stack1.dot', 2, 1, TRUE, TRUE, TRUE);
+-------------------+
| summary           |
+-------------------+
| Disabled 1 thread |
+-------------------+
1 row in set (0.05 sec)

+---------------------------------------------+
| Info                                        |
+---------------------------------------------+
| Data collection starting for THREAD_ID = 75 |
+---------------------------------------------+
1 row in set (0.10 sec)

+-------------------------------------------------------+
| Error                                                 |
+-------------------------------------------------------+
| File '/var/lib/mysql-files/stack1.dot' already exists |
+-------------------------------------------------------+
1 row in set (2.11 sec)

+------------------+
| summary          |
+------------------+
| Enabled 1 thread |
+------------------+
1 row in set (2.31 sec)

Query OK, 0 rows affected (2.56 sec)

mysql> SELECT COUNT(*) FROM performance_schema.setup_actors;
+----------+
| COUNT(*) |
+----------+
|        2 |
+----------+
1 row in set (0.05 sec)

mysql> SELECT COUNT(*) FROM performance_schema.threads WHERE INSTRUMENTED = 'YES';
+----------+
| COUNT(*) |
+----------+
|       33 |
+----------+
1 row in set (0.05 sec)
```